tenant related components moved to tenant folder
